converted test to factory
